+++ src/main/java/com/klikli_dev/occultism/config/OccultismServerConfig.java
@@ -31,7 +31,6 @@ public class OccultismServerConfig {
     public final SpiritJobSettings spiritJobs;
 
     public final RitualSettings rituals;
-    public final DimensionalMineshaftSettings dimensionalMineshaft;
     public final ItemSettings itemSettings;
     public final ModConfigSpec spec;
 
@@ -40,7 +39,6 @@ public class OccultismServerConfig {
         this.storage = new StorageSettings(builder);
         this.spiritJobs = new SpiritJobSettings(builder);
         this.rituals = new RitualSettings(builder);
-        this.dimensionalMineshaft = new DimensionalMineshaftSettings(builder);
         this.itemSettings = new ItemSettings(builder);
         this.spec = builder.build();
     }
@@ -135,55 +133,6 @@ public class OccultismServerConfig {
 
             builder.pop();
         }
-    }
-
-    public static class DimensionalMineshaftSettings {
-        public final MinerSpiritSettings minerFoliotUnspecialized;
-        public final MinerSpiritSettings minerDjinniOres;
-        public final MinerSpiritSettings minerAfritDeeps;
-        public final MinerSpiritSettings minerMaridMaster;
-
-        public DimensionalMineshaftSettings(ModConfigSpec.Builder builder) {
-            builder.comment("Dimensional Mineshaft Settings").push("dimensional_mineshaft");
-
-            this.minerFoliotUnspecialized =
-                    new MinerSpiritSettings("miner_foliot_unspecialized", builder, 400, 1, 1000);
-
-            this.minerDjinniOres =
-                    new MinerSpiritSettings("miner_djinni_ores", builder, 300, 1, 400);
-
-            this.minerAfritDeeps =
-                    new MinerSpiritSettings("miner_afrit_deeps", builder, 200, 1, 800);
-
-            this.minerMaridMaster =
-                    new MinerSpiritSettings("miner_marid_master", builder, 100, 1, 1600);
-
-            builder.pop();
-        }
-
-        public static class MinerSpiritSettings {
-            public final IntValue maxMiningTime;
-            public final IntValue rollsPerOperation;
-            public final IntValue durability;
-
-            public MinerSpiritSettings(String oreName, ModConfigSpec.Builder builder,
-                                       int maxMiningTime, int rollsPerOperation, int durability) {
-                builder.comment("Miner Spirit Settings").push(oreName);
-
-                this.maxMiningTime =
-                        builder.comment("The amount of time it takes the spirit to perform one mining operation.")
-                                .defineInRange("maxMiningTime", maxMiningTime, 0, Integer.MAX_VALUE);
-                this.rollsPerOperation =
-                        builder.comment("The amount of blocks the spirit will obtain per mining operation")
-                                .defineInRange("rollsPerOperation", rollsPerOperation, 0, Integer.MAX_VALUE);
-                this.durability =
-                        builder.comment("The amount of mining operations the spirit can perform before breaking.")
-                                .defineInRange("durability", durability, 0, Integer.MAX_VALUE);
-
-                builder.pop();
-            }
-        }
-
     }
 
     public static class RitualSettings {